eliminate circular dependency
